face_features_app/requirements.txt
tensorflow
keras-cv
gradio
opencv-python
numpy
requests 
Pillow
